refactor(website): various config grids
