+++ apps/website/src/app/dashboard/[id]/modmail/categories/new/_components/CreateCategoryForm.tsx
@@ -1,0 +1,275 @@
+'use client';
+
+import { createCategoryBodySchema } from '@chatsift/api/modmail-schemas';
+import Link from 'next/link';
+import { useParams, useRouter } from 'next/navigation';
+import { useState } from 'react';
+import { APIError } from '@/api/error';
+import { useGuildInfo } from '@/api/routes/guilds';
+import type { CreateModmailCategoryBody } from '@/api/routes/modmail';
+import {
+	useCreateModmailCategory,
+	useModForumTags,
+	useModmailCategories,
+	useModmailConfig,
+} from '@/api/routes/modmail';
+import { Button } from '@/components/common/Button';
+import { EmojiInput } from '@/components/common/EmojiInput';
+import { ForumTagSelect } from '@/components/common/ForumTagSelect';
+import { TemplatePlaceholdersHint } from '@/components/common/TemplatePlaceholdersHint';
+
+interface CategoryFormData {
+	description: string;
+	emoji: string;
+	forumTagId: string;
+	greetingMessage: string;
+	maxConcurrentThreads: string;
+	name: string;
+}
+
+type CategoryFormErrors = Partial<Record<keyof CategoryFormData, string>>;
+
+const EMPTY_FORM: CategoryFormData = {
+	name: '',
+	emoji: '',
+	description: '',
+	greetingMessage: '',
+	forumTagId: '',
+	maxConcurrentThreads: '',
+};
+
+const CATEGORY_FIELDS = [
+	'name',
+	'emoji',
+	'description',
+	'greetingMessage',
+	'forumTagId',
+	'maxConcurrentThreads',
+] as const satisfies (keyof CategoryFormData)[];
+
+function mapCategoryIssues(issues: readonly { message: string; path: PropertyKey[] }[]): CategoryFormErrors {
+	const errors: CategoryFormErrors = {};
+
+	for (const issue of issues) {
+		const [first] = issue.path;
+		if (typeof first === 'string' && (CATEGORY_FIELDS as readonly string[]).includes(first)) {
+			errors[first as keyof CategoryFormData] ??= issue.message;
+		}
+	}
+
+	return errors;
+}
+
+export function CreateCategoryForm() {
+	const router = useRouter();
+	const { id: guildId } = useParams<{ id: string }>();
+
+	const [form, setForm] = useState<CategoryFormData>(EMPTY_FORM);
+	const [errors, setErrors] = useState<CategoryFormErrors>({});
+	const createCategory = useCreateModmailCategory(guildId);
+	const { data: categories } = useModmailCategories(guildId);
+	const { data: guildInfo } = useGuildInfo(guildId, 'MODMAIL');
+	const { data: config } = useModmailConfig(guildId);
+	const { tags: forumTags, modForumConfigured } = useModForumTags(guildId);
+
+	const updateField = (field: keyof CategoryFormData, value: string) => {
+		setForm((prev) => ({ ...prev, [field]: value }));
+		setErrors((prev) => ({ ...prev, [field]: undefined }));
+	};
+
+	const handleSubmit = async (e: React.FormEvent) => {
+		e.preventDefault();
+
+		const data: CreateModmailCategoryBody = {
+			name: form.name.trim(),
+			emoji: form.emoji.trim() || null,
+			description: form.description.trim() || null,
+			greetingMessage: form.greetingMessage.trim() || null,
+			forumTagId: form.forumTagId.trim() || null,
+			maxConcurrentThreads: form.maxConcurrentThreads.trim() ? Number(form.maxConcurrentThreads) : null,
+			// New categories always append at the end -- reordering afterwards happens via the move up/down
+			// controls on each card, not by hand-picking a number here.
+			sortOrder: categories?.length ?? 0,
+		};
+
+		const result = createCategoryBodySchema.safeParse(data);
+		if (!result.success) {
+			setErrors(mapCategoryIssues(result.error.issues));
+			return;
+		}
+
+		try {
+			await createCategory.mutateAsync(result.data);
+			router.replace(`/dashboard/${guildId}/modmail/categories`);
+		} catch (error) {
+			if (error instanceof APIError) {
+				if (error.conflictField) {
+					// `conflictField` is a structured indicator the API attaches to a domain error outside plain
+					// zod validation -- either a 409 (duplicate name, duplicate forum tag) or the 400 this route
+					// throws when `maxConcurrentThreads` exceeds the guild's general limit (see
+					// createCategory.ts/sendBoom.ts). Checked before the zod-validation branch below since
+					// those errors carry no `validationErrors` tree for `fieldError` to read. Falls back to
+					// `name` only as defense-in-depth against a future conflict this route hasn't set one for.
+					const field = (CATEGORY_FIELDS as readonly string[]).includes(error.conflictField)
+						? (error.conflictField as keyof CategoryFormData)
+						: 'name';
+					setErrors({ [field]: error.message });
+				} else if (error.statusCode === 400) {
+					setErrors(
+						Object.fromEntries(
+							CATEGORY_FIELDS.map((field) => [field, error.fieldError(field)]).filter(([, message]) => message),
+						),
+					);
+				} else {
+					setErrors({ name: error.message || 'Failed to create category' });
+				}
+
+				return;
+			}
+
+			setErrors({ name: 'Failed to create category' });
+			console.error('Failed to create category', error);
+		}
+	};
+
+	return (
+		<form className="mt-8 space-y-6" onSubmit={handleSubmit}>
+			<div className="space-y-4">
+				<h2 className="text-xl font-medium text-primary dark:text-primary-dark">Category Details</h2>
+
+				<div>
+					<label
+						className="mb-1 block text-sm font-medium text-secondary dark:text-secondary-dark"
+						htmlFor="category-name"
+					>
+						Name *
+					</label>
+					<input
+						className="w-full rounded-md border border-on-secondary bg-card px-3 py-2 text-primary focus:border-misc-accent focus:outline-none focus:ring-2 focus:ring-misc-accent dark:border-on-secondary-dark dark:bg-card-dark dark:text-primary-dark"
+						id="category-name"
+						maxLength={100}
+						onChange={(e) => updateField('name', e.target.value)}
+						placeholder="e.g. Report a user"
+						type="text"
+						value={form.name}
+					/>
+					{errors.name && <p className="mt-1 text-sm text-misc-danger">{errors.name}</p>}
+				</div>
+
+				<EmojiInput
+					emojis={guildInfo?.emojis ?? []}
+					error={errors.emoji}
+					id="category-emoji"
+					label="Emoji"
+					onChange={(value) => updateField('emoji', value)}
+					placeholder="🚩"
+					value={form.emoji}
+				/>
+
+				<div>
+					<label
+						className="mb-1 block text-sm font-medium text-secondary dark:text-secondary-dark"
+						htmlFor="category-description"
+					>
+						Description
+					</label>
+					<textarea
+						className="w-full rounded-md border border-on-secondary bg-card px-3 py-2 text-primary focus:border-misc-accent focus:outline-none focus:ring-2 focus:ring-misc-accent dark:border-on-secondary-dark dark:bg-card-dark dark:text-primary-dark"
+						id="category-description"
+						maxLength={500}
+						onChange={(e) => updateField('description', e.target.value)}
+						rows={2}
+						value={form.description}
+					/>
+					{errors.description && <p className="mt-1 text-sm text-misc-danger">{errors.description}</p>}
+				</div>
+
+				<div>
+					<label
+						className="mb-1 block text-sm font-medium text-secondary dark:text-secondary-dark"
+						htmlFor="category-greeting"
+					>
+						Greeting Message
+					</label>
+					<textarea
+						className="w-full rounded-md border border-on-secondary bg-card px-3 py-2 text-primary focus:border-misc-accent focus:outline-none focus:ring-2 focus:ring-misc-accent dark:border-on-secondary-dark dark:bg-card-dark dark:text-primary-dark"
+						id="category-greeting"
+						maxLength={2_000}
+						onChange={(e) => updateField('greetingMessage', e.target.value)}
+						rows={2}
+						value={form.greetingMessage}
+					/>
+					<p className="mt-1 text-sm text-secondary dark:text-secondary-dark">
+						Falls back to the{' '}
+						<Link className="underline hover:text-misc-accent" href={`/dashboard/${guildId}/modmail/config`}>
+							guild default
+						</Link>{' '}
+						if unset.
+					</p>
+					<TemplatePlaceholdersHint />
+					{errors.greetingMessage && <p className="mt-1 text-sm text-misc-danger">{errors.greetingMessage}</p>}
+				</div>
+
+				{modForumConfigured ? (
+					<ForumTagSelect
+						error={errors.forumTagId}
+						id="category-forum-tag"
+						label="Forum Tag"
+						onChange={(value) => updateField('forumTagId', value ?? '')}
+						tags={forumTags ?? []}
+						value={form.forumTagId}
+					/>
+				) : (
+					<p className="text-sm text-secondary dark:text-secondary-dark">
+						No Mod Forum configured — set one on the Config page to route this category to a forum tag.
+					</p>
+				)}
+
+				<div>
+					<label
+						className="mb-1 block text-sm font-medium text-secondary dark:text-secondary-dark"
+						htmlFor="category-max-concurrent-threads"
+					>
+						Max Concurrent Threads Override
+					</label>
+					<input
+						className="w-full rounded-md border border-on-secondary bg-card px-3 py-2 text-primary focus:border-misc-accent focus:outline-none focus:ring-2 focus:ring-misc-accent dark:border-on-secondary-dark dark:bg-card-dark dark:text-primary-dark"
+						id="category-max-concurrent-threads"
+						min={1}
+						onChange={(e) => updateField('maxConcurrentThreads', e.target.value)}
+						placeholder={config ? String(config.maxConcurrentThreads) : ''}
+						type="number"
+						value={form.maxConcurrentThreads}
+					/>
+					<p className="mt-1 text-sm text-secondary dark:text-secondary-dark">
+						How many tickets a user may have open in this category specifically. Leave blank to use the{' '}
+						<Link className="underline hover:text-misc-accent" href={`/dashboard/${guildId}/modmail/config`}>
+							guild default
+						</Link>
+						{config ? ` (currently ${config.maxConcurrentThreads})` : ''}. Cannot exceed the guild default.
+					</p>
+					{errors.maxConcurrentThreads && (
+						<p className="mt-1 text-sm text-misc-danger">{errors.maxConcurrentThreads}</p>
+					)}
+				</div>
+			</div>
+
+			<div className="flex gap-4">
+				<Button
+					className="px-3 py-2.5 bg-misc-accent text-white rounded-md hover:opacity-90 transition-opacity disabled:opacity-50 disabled:cursor-not-allowed"
+					isDisabled={!form.name.trim() || createCategory.isPending}
+					type="submit"
+				>
+					{createCategory.isPending ? 'Creating...' : 'Add Category'}
+				</Button>
+				<Button
+					className="px-3 py-2.5 bg-on-tertiary dark:bg-on-tertiary-dark text-primary dark:text-primary-dark rounded-md hover:bg-on-secondary dark:hover:bg-on-secondary-dark transition-colors"
+					onPress={() => router.back()}
+					type="button"
+				>
+					Cancel
+				</Button>
+			</div>
+		</form>
+	);
+}
